TS30DEV-16 [수정] 연결 내용 수정 오동작 수정

+++ pkg/models-thingspin/fmsConnect.go
@@ -56,7 +56,7 @@ type UpdateTsConnectFlowQuery struct {
 	Enable    bool
 	Intervals int64
 	Params    string
-	Result    int64
+	Result    int
 }
 
 type UpdateActiveTsConnectQuery struct {
